requirements.txt
absl-py==2.1.0
accelerate==1.2.0
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
audioread==3.0.1
av==9.2.0
basicsr==1.4.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
diffusers==0.31.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
filterpy==1.4.5
fire==0.7.0
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gfpgan==1.3.8
gradio==5.8.0
gradio_client==1.5.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
jax==0.4.37
jaxlib==0.4.36
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lmdb==1.5.1
login==0.0.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==3.20.3
psutil==6.1.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
stable_diffusion_videos==0.9.1
starlette==0.41.3
sympy==1.13.1
tb-nightly==2.19.0a20241211
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
